geodata/geodjango/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cities==0.6.1
django-model-utils==4.3.1
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
tqdm==4.66.1
typing_extensions==4.7.1
